Crate kestrel_runtime

The KILT runtime. This can be compiled with #[no_std], ready for Wasm.

Re-exports

pub use attestation;

pub use ctype;

pub use delegation;

pub use did;

pub use pallet_web3_names;

pub use public_credentials;

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AttestationDeposit

AutoUnlockBound

BlockHashCount

DelegationDeposit

DidFee

ExistentialDeposit

Fee

MaxAuthorities

MaxBlocksTxValidity

MaxChildren

MaxClaims

MaxDelegatedAttestations

MaxFreezes

MaxHolds

MaxLocks

MaxNewKeyAgreementKeys

MaxNominators

MaxNumberOfServicesPerDid

MaxNumberOfTypesPerService

MaxNumberOfUrlsPerService

MaxParentChecks

MaxPublicKeysPerDid

MaxRemovals

MaxReserves

MaxRevocations

MaxServiceIdLength

MaxServiceTypeLength

MaxServiceUrlLength

MaxSignatureByteLength

MaxTotalKeyAgreementKeys

MinimumPeriod

Offset

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Period

Permill

A fixed point representation of a number in the range [0, 1].

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

UncleGenerations

UsableBalance

Version

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

TimestampCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Attestation

Aura

AuraConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

Ctype

Delegation

Did

DidLookup

DidLookupConfig

DigestItem

Digest item type.

Executive

Executive: handles dispatch to the various Pallets.

GenesisConfigDeprecated

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Indices

IndicesConfig

Multisig

Proxy

PublicCredentials

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Web3Names